Not sure what I changed...not much for sure.
I think I have a way to improve the run time quite a bit. Since many
different HNFs (of the same SNF) have the list of rotation
permutations, we should loop over those instead of over the HNFs. For
each class, you only need one list of labelings. The potential savings
a factor of 4 or 5.