The last problem is indeed a problem with the compiler. No problem on the Intel compiler.
I found a bug where the identity permutation was not being generated
for some HNFs. The problem was that the index for a rotation in the d-set
was not correct. I had to keep track of which rotations survived for each 
HNF and make an index so I could map the group members correctly. It was really
messy. I had to pass the RPList in and out of several routines so that
it would be properly initialized and rearranged with the SNFs and HNFs.