I'm trying to change the enumeration code to allow different sets of labels on different points in the multilattice. That is, allow for things like the following: Two points in the parent cell, one is labeled 0 or 1, the second is labeled 2 or 3.
In this revision, I've written a short module, mixed_radix_counter, to test the idea of
generating labelings that follow this idea. Then I changed the read_input routine to read
in a new format of struct_enum.in file. This file has a list of labels for each d-vectors,
and the k input (binary, ternary,etc.) now refers to the entire number of labels (can be
different numbers of labels for different sites).

The new input is read in but *not* used yet in this revision. The next step is to pass out the 
information, pass it in to the appropriate routines, and to change the label generator (in the
end of the generate_unique_labelings routine labeling_related.f90.