There was a bug in makeStr.f90 and makeStrIn.f90. (The latter still remains.)
This caused makestr.x to write POSCARs that were incorrect. The cell and 
atomic sites were correct but the labeling (atom types) were not.

The problem was that the routine that generates the real space coordinates
of the atoms (see interior_points.pdf) from the info in the
struct_enum.out file doesn't generate the atoms in the *same order* as
the enumeration code does when it writes the labeling to
struct_enum.out. Indeed there's no reason to expect it to, but I
overlooked this when writing makeStr.f90 and makeStrIn.f90. I was just applying
the labels in the labeling to the atoms in the order they came out.

I have fixed makeStr.f90 (the L matrix is important---don't forget to transpose after
reading it in!) I'm not going to fix makeStrIn.f90 until I have a single
routine for doing the conversion that I can use in both programs.