Just added a comment to makeStr.f90
This routine also now echoes the input that it reads to a file. It also
writes out the computations that it does to the same file. This is great
for debugging when the struct_enum.out format changes or when the input
is somehow wrong. All our routines should do this.