Updated the tests for the compare code (still needs some work though).
Also changed two things in the enumeration itself. One was a bad failsafe 
test for the new mixed-radix approach to generating labelings. This 
failsafe was causing the code to quit when it shouldn't (two sites, one 
labeled 0/1, the other 0/2) caused problems but  shouldn't have. Also, I 
changed the "idx=" line in labeling related. It now is computed based of 
digCnt instead of the labeling "a". I'm a little leary that there are still
bugs in the enumeration code with the new mixed-radix stuff.

Now that the compare code is working, it's time to build a test and see if
the enumeration code still gives the same answers as it use to. This is 
a high priority.