Removed makeStrIn.f90 (replaced by make_structures.in.py).
Updated the campare tests.

Changed the way idx is computed in labeling_related (this was
required by the mixed-radix stuff)

Commented out the new line in struct_enum.out. Add this back
in Tobias if you need it but we need to see what other
routines that read in the header of this file...