Found a shortcoming in the find_structure_in_list program when trying to find the D0_19 structure in a list of hcp enumerations. The D0_19 poscar was generated by aconvasp and placed the inversion center of the structure at the origin (rather than one of the lattice sites like I normally do it). This causeh the program to fail because the lattice sites between the enum file and D0_19 structure were not coincident.
Changed the program (mainly enumeration_utilities) so that the parent lattice sites
in the enum file and parent lattice sites in POSCAR are consistent (they don't need to match---can
be an origin shift). The origin shift is computed and then the lattice sites of the check structure
are shifted.

Changed the makefile too (removed the -xW option for "ifort". This isn't a valid flag
for my compiler).

Also fixed a bug in get_sLV_fixing_operations---tmpOp%rot and tmpOp%shift were not allocated.

Also changed the glob command in comparetests.py. Some OS's alphabetize, others don't. Need to
guarantee order for the 17test to pass.

I also changed some of the input for the tests---order of things in struct_enum.out had changed...

Also added some tests to comparetests.py