Added Rod's write-up of the "multiperms" algorithm to the repository. (It's also in the repository for the enumIII paper.)
Added a routine to io_utils that reads a list of unit cells from a file called
"fixed_cells.in" The allows us to search over configurations for fixed cells 
(rather than enumerating cells) if we want to. If "fixed_cells.in" is present, 
multienum.x (and derivative_structure_generator) will skip the HNF generation and
just use the cells in the file.

struct_enum.in has an extra line now so that we can specify a concentration range 
(as a list of rational numbers, last number is the denominator). This functionality
needs to be extended to k-nary cases (only binary is assumed now).

derivative_structure generator has been minimally changed in several places to allow
for the fixed cell list and for a specified concentration range.

labeling_related now has the extra routine, generate_permutation_labelings. It's not done
yet (just the beginnings are there) but I want to commit so I can backup my changes so
far.

All changes here compile and work fine already except the labeling_related routine
generate_permutation_labelings.