Just changed a couple of declaration statements for some pointers in labeling_related.f90
There was a strange bug where the HNF generator was generating HNFs that weren't in HNF
form. Lance and I can't remember if we fixed it, but we can't reproduce it now. It seems
that it was a problem with a pointer (hence the changes above) but we can't remember
or get the error to occur again.