Cleaned up old files in the main directory (mainly old input files and things that were incorrectly added to the repo).
Removed the call to get_spaceGroup_atomTypes(label,digit,aTyp) in
derivative_structure_generator. This reduced the symmetry incorrectly
for sites with different labelings. However, in order to treat Tobias's
surface problems (where two seemingly equivalent sites are different
because of their environment, we need to generalize the input to
have an optional "environment tag" for each site.) This could be added 
just after the list of labels for each d-vector