Fixed a design shortcoming in labeling_related.f90 that was exposed when Tobias and Bryce started doing complicated multi-lattices.
The problem was that the counter for the number of digits was
initialized assuming that every label appeared on every
site. Obviously this is not true if some labels don't ever appear on
some site. I'm surprised that this didn't come up sooner, frankly.

I also had to comment out the deallocate for "lab" Not sure why that
is a problem. "associated" never seems to work right with
ifort. Hopefully it isn't a harbinger of some super-subtle bug that's
my fault...