The compare code is not working at this stage, but the changes are underway and I'm going to commit so that I have a checkpoint.
Changes were necessary to this code when the SNF algorithm in 
rational_mathematics changed. The change was made so that the SNF code 
would be more robust against overflow, but because the L and R matrices 
are not unique, this changed the output for enumlib (old runs and new 
runs have equivalent but not identical output).

We want the compare code to continue working because it's central to the 
unit tests. But the changes in the L matrices mean that the labelings 
can be different and this makes it difficult to compare old runs and new 
runs (before and after the SNF algorith change.)

Rod and I had to remember a lot of enumlib-related group theory and 
re-read the first two enumeration papers to sort out what we needed to 
do to determine equivalence between two different enumerations 
(labelings) that should be equivalent but were not identical.

For now, the details of that effort are in Gus' email, as documented in 
the comments in this file. When it's all sorted out and working, 
probably should add a write up to document the group theory and the 
algorithms in the code.