The compare code isn't working correctly yet, but I'm committing so that I have a checkpoint.
Renamed `ilabeling` to `ilabeling1` (parallelism with `ilabeling2`

Added the code to handle self-complementary equiatomic cases (but it has 
not been debugged yet because the basic structure of the search is 
currently flawed.)

Right now, the compare code (due to flawed attempts to fix it before Rod 
and GLWH figured out the right way to handle the equiatoimic case, the 
code incorrectly loops over permutations of the label in *both* file1 
and file2. This is incorrect. This is the what the next commit hopefully 
fixes. This commit is just to give a checkpoint before some serious 
surgery.